+++ src/pages/mission/components/mission/Challenge.tsx
